requirements.txt
py-cord==2.4.1
python-dotenv==1.0.0